Reworked "Prop Bonus Round" into "Propify!"
Functionality of the "Prop Bonus Round" plugin has been moved to
"Propify!", a standalone plugin that allows player propping by admins
and exposes propping methods; and a new "Prop Bonus Round" that uses the
exposed methods from "Propify!" to force the losing team into props.

Though this means the web compiler cannot be used to compile the
end-round functionality of the plugin, it also opens up the ability of
propping players to be integrated elsewhere and functionality is fully
modularized.